fix Keepright task
