+++ src/org/openstreetmap/josm/plugins/tofix/bean/items/ItemKeeprightBean.java
@@ -1,5 +1,9 @@
 package org.openstreetmap.josm.plugins.tofix.bean.items;
 
+import java.util.Arrays;
+import org.openstreetmap.josm.data.coor.LatLon;
+import org.openstreetmap.josm.data.osm.Node;
+
 /**
  *
  * @author ruben
@@ -7,7 +11,7 @@ package org.openstreetmap.josm.plugins.tofix.bean.items;
 public class ItemKeeprightBean {
 
     private String key;
-    private ItemKeeprightValueBean value;
+    private Value value;
 
     public String getKey() {
         return key;
@@ -17,12 +21,53 @@ public class ItemKeeprightBean {
         this.key = key;
     }
 
-    public ItemKeeprightValueBean getValue() {
+    public Value getValue() {
         return value;
     }
 
-    public void setValue(ItemKeeprightValueBean value) {
+    public void setValue(Value value) {
         this.value = value;
     }
-    
+
+    public class Value {
+
+        String object_type;
+        Long object_id;
+        String st_astext;
+
+        public String getObject_type() {
+            return object_type;
+        }
+
+        public void setObject_type(String object_type) {
+            this.object_type = object_type;
+        }
+
+        public Long getObject_id() {
+            return object_id;
+        }
+
+        public void setObject_id(Long object_id) {
+            this.object_id = object_id;
+        }
+
+        public String getSt_astext() {
+            return st_astext;
+        }
+
+        public void setSt_astext(String st_astext) {
+            this.st_astext = st_astext;
+        }
+
+        public Node get_node() {
+            String str = getSt_astext().replaceAll("[^-?0-9.]+", " ");
+            Double lat = Double.parseDouble(Arrays.asList(str.trim().split(" ")).get(1));
+            Double lon = Double.parseDouble(Arrays.asList(str.trim().split(" ")).get(0));
+            LatLon coor = new LatLon(lat, lon);
+            Node node = new Node(coor);
+            return node;
+        }
+
+    }
+
 }
